# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My Doc API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/users": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Список",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Some name"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Создание",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "firstname": {
                                                "type": "string",
                                                "example": "Some firstname"
                                            },
                                            "lastname": {
                                                "type": "string",
                                                "example": "Some lastname"
                                            },
                                            "address": {
                                                "type": "string",
                                                "example": "Some address"
                                            },
                                            "email": {
                                                "type": "string",
                                                "example": "user@email.com"
                                            },
                                            "password": {
                                                "type": "string",
                                                "example": "Some password"
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": "1"
                                                },
                                                "firstname": {
                                                    "type": "string",
                                                    "example": "Some firstname"
                                                },
                                                "lastname": {
                                                    "type": "string",
                                                    "example": "Some lastname"
                                                },
                                                "address": {
                                                    "type": "string",
                                                    "example": "Some address"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "user@email.com"
                                                },
                                                "password": {
                                                    "type": "string",
                                                    "example": "Some password"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "options": {
                "tags": [
                    "User"
                ],
                "summary": "Получение доступных методов",
                "responses": {
                    "200": {
                        "description": "Ok"
                    }
                }
            },
            "head": {
                "tags": [
                    "User"
                ],
                "summary": "Проверка существования ресурса",
                "responses": {
                    "200": {
                        "description": "Ok"
                    }
                }
            }
        },
        "/api/users/{user}": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Пользователь",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "Id пользователя",
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": "1"
                                                },
                                                "firstname": {
                                                    "type": "string",
                                                    "example": "Some firstname"
                                                },
                                                "lastname": {
                                                    "type": "string",
                                                    "example": "Some lastname"
                                                },
                                                "address": {
                                                    "type": "string",
                                                    "example": "Some address"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "user@email.com"
                                                },
                                                "password": {
                                                    "type": "string",
                                                    "example": "Some password"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "User"
                ],
                "summary": "Удаление",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "Id пользователя",
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "integer",
                                            "example": "done"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "User"
                ],
                "summary": "Обновление",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "Id пользователя",
                        "required": true,
                        "example": 1
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "firstname": {
                                                "type": "string",
                                                "example": "Some firstname"
                                            },
                                            "lastname": {
                                                "type": "string",
                                                "example": "Some lastname"
                                            },
                                            "address": {
                                                "type": "string",
                                                "example": "Some address"
                                            },
                                            "email": {
                                                "type": "string",
                                                "example": "user@email.com"
                                            },
                                            "password": {
                                                "type": "string",
                                                "example": "Some password"
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": "1"
                                                },
                                                "firstname": {
                                                    "type": "string",
                                                    "example": "Some firstname"
                                                },
                                                "lastname": {
                                                    "type": "string",
                                                    "example": "Some lastname"
                                                },
                                                "address": {
                                                    "type": "string",
                                                    "example": "Some address"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "user@email.com"
                                                },
                                                "password": {
                                                    "type": "string",
                                                    "example": "Some password"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}